+++ client/src/App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useContext, useEffect } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import Navbar from './components/layout/Navbar';
@@ -7,34 +7,41 @@ import Register from './components/auth/Register';
 import Login from './components/auth/Login';
 import Alert from './components/layout/Alert';
 
-import { AlertProvider } from './AlertContext';
-import { UserProvider } from './UserContext';
+import { UserContext } from './UserContext';
+import setAuthToken from './utils/setAuthToken';
 
 import './App.css';
 
-const App = () => (
-    <AlertProvider>
-        <UserProvider>
-            <Router>
-                <>
-                    <Navbar />
-                    <Route exact path='/'>
-                        <Landing />
-                    </Route>
-                    <section className='container'>
-                        <Alert />
-                        <Switch>
-                            <Route exact path='/register'>
-                                <Register />
-                            </Route>
-                            <Route exact path='/login'>
-                                <Login />
-                            </Route>
-                        </Switch>
-                    </section>
-                </>
-            </Router>
-        </UserProvider>
-    </AlertProvider>
-);
+if (localStorage.token) {
+    setAuthToken(localStorage.token);
+}
+
+const App = () => {
+    const { loadUser } = useContext(UserContext);
+    useEffect(() => {
+        loadUser();
+    }, []);
+
+    return (
+        <Router>
+            <>
+                <Navbar />
+                <Route exact path='/'>
+                    <Landing />
+                </Route>
+                <section className='container'>
+                    <Alert />
+                    <Switch>
+                        <Route exact path='/register'>
+                            <Register />
+                        </Route>
+                        <Route exact path='/login'>
+                            <Login />
+                        </Route>
+                    </Switch>
+                </section>
+            </>
+        </Router>
+    );
+};
 export default App;