+++ client/src/UserContext.js
@@ -2,6 +2,7 @@ import React, { useState, useContext } from 'react';
 import axios from 'axios';
 
 import { AlertContext } from './AlertContext';
+import setAuthToken from './utils/setAuthToken';
 
 const UserContext = React.createContext();
 
@@ -14,6 +15,42 @@ function UserProvider({ children }) {
     });
 
     const { setAlert } = useContext(AlertContext);
+
+    async function loadUser() {
+        if (localStorage.token) {
+            setAuthToken(localStorage.token);
+        }
+        try {
+            const res = await axios.get('/api/auth');
+            setUser(prevState => ({
+                ...prevState,
+                isAuthenticated: true,
+                loading: false,
+                user: res
+            }));
+        } catch (err) {
+            const errors = err.response.data.errors;
+            authError(errors);
+        }
+    }
+
+    async function login(email, password) {
+        const config = {
+            headers: {
+                'Content-Type': 'application/json'
+            }
+        };
+        const body = JSON.stringify({ email, password });
+        try {
+            const res = await axios.post('/api/auth', body, config);
+            localStorage.setItem('token', res.data.token);
+            loadUser();
+            setAlert('Welcome back!', 'success');
+        } catch (err) {
+            const errors = err.response.data.errors;
+            authError(errors);
+        }
+    }
 
     async function register(name, email, password) {
         const config = {
@@ -28,7 +65,7 @@ function UserProvider({ children }) {
             setAlert('You have successfully signed up!', 'success');
         } catch (err) {
             const errors = err.response.data.errors;
-            registerFail(errors);
+            authError(errors);
         }
     }
 
@@ -42,7 +79,7 @@ function UserProvider({ children }) {
         }));
     }
 
-    function registerFail(errors) {
+    function authError(errors) {
         localStorage.removeItem('token');
         setUser(prevState => ({
             ...prevState,
@@ -57,7 +94,14 @@ function UserProvider({ children }) {
 
     return (
         <UserContext.Provider
-            value={{ userObj, register, registerSuccess, registerFail }}
+            value={{
+                userObj,
+                register,
+                registerSuccess,
+                authError,
+                loadUser,
+                login
+            }}
         >
             {children}
         </UserContext.Provider>